==> docs/03-geocoding-gateway.Rmd <==
---
title: "Test Gateway's Geocoder"
author: "Branson Fox"
date: '(`r format(Sys.time(), "%B %d, %Y")`)'
output: 
  github_document: default
  html_notebook: default
---

## Introduction

## Dependencies
These packages are required for the analysis in this notebook.

```{r dependencies}
# tidystl
library(dplyr)
library(gateway)
library(stringr)

# tidyverse packages
library(readr)

# other packages
library(here)
```


## Load Data
Next, we'll load the data to be geocoded:

```{r load-data}
addresses <- read_csv(here("data", "STL_CRIME_Homicides.csv"))
```

## Build Geocoders

```{r}
geocoder <- gw_build_geocoder(style = "full", return = "coords")
geocoder_s <- gw_build_geocoder(style = "short", return = "coords")
geocoder_p <- gw_build_geocoder(style = "placename", return = "coords")
```

```{r}
# remove dashes
geocoder_p <- mutate(geocoder_p, gw_name = str_replace_all(string = gw_name, pattern = "-", replacement = " "))
```


```{r}
gateway_time <- system.time({
  gateway <- gw_geocode(addresses, type = "composite, full",
                        var = address_norm, class = "tibble",
                        local = geocoder, local_short = geocoder_s,
                        local_place = geocoder_p, threshold = 90)
})

save(gateway, gateway_time, file = here("results", "gateway.rda"))
```
