==> docs/05-census-tracts.Rmd <==
---
title: "Tract-Level Accuracy"
author: "Branson Fox"
date: '(`r format(Sys.time(), "%B %d, %Y")`)'
output: 
  github_document: default
  html_notebook: default
---

## Objective
This notebooks is used to interpolate our geocoding results to census tracts, a common unit of aggregate spatial analysis. From this, we will evaluate the quality of geocoding as it pertains to analysis at the census tract level.

## Dependencies
The following packages are necessary to replicate this process.
```{r dependencies, message=FALSE}
library(sf)
library(dplyr)
library(magrittr)
library(tigris)
library(tidycensus)
library(areal)
```

## Load Data
We'll load our saved geocoding results. We will also filter to points where we have a valid ground truth. We'll also replace NAs with 0s in order to make valid SF objects. These won't match any tracts.
```{r load data}
load('../results/all_results.rda')
all_geocodes %<>% filter(!is.na(truth.lat))
all_geocodes[is.na(all_geocodes)] <- 0
```

We must also load the census tracts for our region of interest.
```{r load tracts}
tracts <- rbind(
  tracts(state = 17, class = 'sf'),
  tracts(state = 29, class = 'sf')
)
stl_tracts <- filter(tracts, COUNTYFP == 510)
tracts %<>% filter(
    st_touches(., stl_tracts, sparse = FALSE) %>% rowSums() > 0
  ) %>% st_transform(4326) %>% transmute(GEOID)
```


## Locate Points in Census Tracts
Now, we must locate each point in the appropriate census tract. We will identify them by the tract they fall within the boundary of.
```{r}
# Define Geocoders
geocoders <- c('truth', 'BingSingle', 'EsriSingle', 'Google', 'HereSingle', 'OpenCage', 'TomTomSingle', 'CensusSingle', 'GeocodioSingle', 'TomTomBatch', 'BingBatch', 'CensusBatch', 'GeocodioBatch', 'EsriBatch', 'HereBatch')

# Initialize a Dataframe to Store Results
located_df <- data.frame(id = 1:nrow(all_geocodes))

# Iterate Through Geocoders, and Store Tract Match in Dataframe
for (geocoder in geocoders){
  # Make an SF Object
  sf <- st_as_sf(all_geocodes, coords = c(
      x = paste0(geocoder, '.lon'),
      y = paste0(geocoder, '.lat')
    ),
    crs = 4326)
  # Calculate Within Predicate
  within <- st_within(sf, tracts)
  # Get GEOID If Exists
  geoid <- lapply(within, function(x){
    if (length(x) < 1){
      return('No Match')
    }else{
      return(tracts[x, 'GEOID'] %>% st_drop_geometry)
    }
  }) %>% unlist %>% unname
  # Append Tract GEOID
  located_df[[geocoder]] <- geoid
  
}
```

## Calculate Exact Matches
We will summarize the number of geocodes that fall within the same tract as the ground truth. We'll operate on the matrix and create a logical matrix
```{r exact matches}
exact <- (located_df[,3:16] == located_df[,2]) %>% colSums %>% sort(decreasing = TRUE)
exact
```


## Calculate Near Matches
Next, we will calculate points that fall within an adjacent tract as the ground truth. To do this, we will first need to define a list of adjacent tracts.
```{r calc adjacent, message=FALSE}
# Get Tract Geometry for Ground Truth
truth_tracts <- left_join(located_df[,1:2], tracts, by = c('truth' = 'GEOID')) %>% st_as_sf()

# Get Adjacent Tracts by Row Number
adjacent <- vector('list', length = nrow(truth_tracts))
for (row in 1:nrow(truth_tracts)){
  adjacent[[row]] <- st_touches(truth_tracts[row,], tracts)
}

# Get GEOID from Row Number
adjacent_geoid <- lapply(adjacent, function(x){
  sapply(x, function(y){
    tracts[y, 'GEOID'] %>% st_drop_geometry()
  }) 
}) %>% unname

# Append to Data.frame
located_df[['adjacent']] <- adjacent_geoid

# Iterate and Append to Data Frame with Results
near_matches <- data.frame(matrix(ncol = 14))
names(near_matches) <- geocoders[2:15]

for (row in 1:nrow(located_df)){
  near_matches %<>% rbind(located_df[row, 3:16] %in% located_df[row, 'adjacent'][[1]]$GEOID)
}
# Remove Blank First Row and Calculate Column Sums
near_matches <- near_matches[2:1517,]
near_matches %<>% colSums
near_matches
```

## Create Shapefiles
In order to conduct a spatial analysis, we must create shapefiles to import into PySal (See the Jupyter Notebooks for the actual analyis)

```{r}
# Get Census Co-Variates
vars <- c(
  'B02001_001', # Race Total
  'B02001_002', # Race White
  'B19049_001', # Median Household Income (2018 Inflation Adjusted)
  'B17020_001', # Total (Poverty) Population
  'B17020_002', # In Poverty
  'B25003_001', # Total (Tenure) Units
  'B25003_002'  # Owner Occupied
)
demog <- get_acs('tract', vars, year = 2018, output = 'wide', state = 29, county = 510, geometry = TRUE) %>%
  transmute(
    non_wht = 1 - B02001_002E / B02001_001E,
    med_inc = B19049_001E,
    poverty = B17020_002E / B17020_001E,
    homeown = B25003_002E / B25003_001E,
    pop = B02001_001E
  ) %>%
  st_transform(26915)

# Then we'll create a process for appending counts of Homicide, and saving shapefiles, and do this for every geocoder including the ground truth
hex <- ar_tessellate(demog, shape = 'hexagon') %>% mutate(tid = row_number())

for (geocoder in geocoders){
  # Extract the Points for the Geocoder and Make an SF Object
  cols <- c(paste0(geocoder, '.lat'), paste0(geocoder, '.lon'))
  sf <- st_as_sf(all_geocodes[,cols], coords = c(x = cols[2], y = cols[1]), crs = 4326) %>% st_transform(26915)
  # Interpolate these Points to Tracts
  agg <- mutate(demog,
          homicides = st_intersects(demog, sf, sparse = FALSE) %>% rowSums() / pop * 1000,
          sid = row_number()
        )
  
  # Interpolate, Using AWI to Hexagons
  awi <- aw_interpolate(hex, tid, agg, sid, 'sum', 'sf', extensive = c('homicides', 'med_inc'), intensive = c('non_wht', 'poverty', 'homeown'))
  
  # Save A Shapefile
  st_write(awi, paste0('../data/shapes/', geocoder, '.shp'))
}

```
